Modify SPIM not to wait for DMA completion in certain cases.
SPIM calls involving the Transfer trait remain unchanged. Also all calls
that involve controlling the CS signal remain unchanged (does the CS
signal really have to stay low while the DMA operation is taking place?).

Callers of the Write trait get the new behavior, as is the case with
display-interface-spi. Better yet, in our case we have a MISO-only
display connection, so we do not have to worry about RAM corruption if
we drop a few compiler fences.